restructure: navbar changes
