+++ src/structures/components/navbar/mobile_navigator.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 
-import Logo                           from '../components/instantiate-logo'
-import PageLinks                      from '../components/instantiate-page-link';
+import Logo                           from '../instantiate-logo'
+import PageLinks                      from '../instantiate-page-link';
 
 
 const MobileNav = () => {
